net: phy: micrel: Don't create ptp clocks if PHY_TIMESTAPING is not enabled
Already in the function lan8814_ptp_init we get out early if the
PHY_TIMESTAMPING is not enabled, but we still create PTP clock in the
function lan8814_ptp_probe_once. There is no function issue for this
only that is confusing to have those clocks when they are not used.
Therefore fix this by checking if the PHY_TIMESTAMPING is enabled and if
it is not just get out without creating clocks.
Because of this change we need to initialize shared->phydev regardless
if PHY_TIMESTAMPING is enabled or not because when an interrupt will
come we will use this, so if it is not initialized we will get an error.