dsa: lan9645x: remove mac table sync via irq
SQA is running a test where 1G line rate traffic cycles through 9000 mac
addresses.

Due the the mac IRQ this puts immense pressure on the CPU. I doubt any
implementation will be able to handle this properly since the CPU is
simply not strong enough.

Apparently MESA does not use the IRQ, but only learns in hardware. The
mac table can be dumped by other means.

The main argument for syncing that I could think of was protocol frames
which are trapped to the CPU. By syncing the mac table, the CPU is able
to forward as the CHIP would do.

However, as Claus mentions, when they are sent out, they are injected
back into the chip, which has all the state necessary to forward it
correctly.

The arguments convinced be to move back to the old fdb_dump routine, and
remove all the mac IRQ code. Now learning at line rate should not impact
the CPU in any way, and there is no failure mode of getting out of sync.