phy: microchip: lan9645x: Fix usage of regmap_update_bits
The function regmap_update_bits has the following declaration:

static inline int regmap_update_bits(struct regmap *map, unsigned int reg,
				     unsigned int mask, unsigned int val);

Where the last argument is the value and the second from last is the
mask but there are 3 usage of this function in this file where the mask and
the val are swapped. Meaning that we don't set necessary the expected value.
Therefore fix this.
We have detected this issue by seeing that the DLL is not getting out of
reset.