dsa: lan9645x: use mactable to control BC to cpu instead of flood masks
The flood masks on ghibli are global and not vlan aware. Either all
BC on all ports in all vlans is forwarded to the CPU, or no BC is.

This is unfortunate, since we want to limit traffic to CPU, but we also
need BC to many protocols. Previously we always had the CPU in the flood
BC mask.

However, inspired by MESA it is possible to use the mac table to get
VLAN aware BC restriction.

The HOST_PVID, UNAWARE_PVID and HSR_PRP reserved vids always get a BC
mac entry, so BC in these VLANs always hit the CPU.

However, for VLAN aware bridged ports we only forward BC to the CPU if
the bridge interface is part of the vlan:

bridge vlan del dev br0 vid 100 self
bridge vlan add dev br0 vid 100 self