dsa: lan9645x: add pcs lost sync recovery
The PCS can lose sync to the serdes, which causes an unrecoverable state
with link flapping.

To recover it is necessary to reset RX in the serdes. This is also seen
by validation, and Marco noticed it during simulation and on the FPGA.

The state can be detected by LINK_STATUS=1 and SYNC=0. This state will
already trigger a link_down/link_up sequence, so we add a bit used at
link down to indicate we must rx reset the serdes.

This is achieved with phy_recalibrate.