net: dsa: lan9645x: fix tc flower ip_ttl validation
At the moment we ignore the ttl field in the ip key, so it is possible
to make a rule with ip_ttl set, but the key is never included.

Instead we should validate if unsupported fields are used in an
otherwise supported key.

The ip key has ttl and tos fields, and we support only tos matching.

The hardware can only match on ttl > 0, but this can not be expressed
in tc.

DBB-973: https://jira.microchip.com/browse/DBB-973