DBB-1003, APPL-6381. uFDMA for MIPS: Short frames were not entirely flushed
Transmission of short frames could cause stale data to appear at the end of the
frames on MIPS platforms. Previously, care had been taken to extend the frame to
a minimum-sized Ethernet frame, but only the original size was flushed to main
memory, so uninitialized data could occur at the end of the frame, causing
Wireshark to show it as a 'Trailer' rather than 'Padding'.