dsa: lan9645x: fix do not use vrule->id as ES0 counter_id for ESDX
The problem with the current logic can overflow the number of available
ESDXs, since we can have more than 128 live vcap rules in total, so
vrule->id > 128.

The ESDX indices are only used by ES0 vcap rules. We attach one per
VCAP rule, because under some conditions, it will give us a u32 counter
instead of the 1bit saturating counter.

At the moment we use vrule->id directly as the ESDX. We know no other
vcap rule has the same ID, and it is convenient when we remove the rule
and need to clear the counter.

However, we only have 128 ESDXs. While this is sufficient to assign one
to every possible ES0 rule, we may have more than 128 vcap rules in
total, in which case vrule->id overflows the number of ESDXs.

To avoid this problem we add some allocation management logic for ESDX
similar to ISDX and SFID.

When we need to add a counter to an ES0 rule we request an ESDX, and
on rule deletion we free it. The freeing logic takes care to clear the
associated counters.