dsa: lan9645x: stats: only update counters for used isdx, esdx and ports
Counters are updated periodically to handle potential for 32bit HW
counters overflowing.

We used to update all counters for all ports, ISDX, ESDX and SFID
indices.

This is more work than necessary, and with register IO over SPI this
represents a significant amount of wasted IO and time.

A rough estimate, at 35Mhz spi frequency, is we go from 200ms per update
to 46ms, when no indices are in use with 7/9 ports in use.