dsa: lan9645x: psfp: integrate with selective stats updating
Only update stats for SFID's which are in use. To achieve this we must
access the sfi_idx_mask from the stat update workqueue. This requires
locking to be added to PSFP for SFI/SGI idx management and table IO.

We simply the addition of the lock by moving the single use of psfp, in
tc flower, to a function lan9645x_psfp_tc_action_set in the unit. This
allows us to just add locks here, and in the two free functions

lan9645x_sgi_put
lan9645x_sfi_put

All now unused functions are taken private and removed from main.h