dsa: lan9645x: hsr: fix npi port rewriter tag_cfg
The blamed commit introduced a bug, because the lan9645x_tag_npi_setup
call was moved earlier in the init process.

The idea was to error early here, before subsystem allocations, if the
npi port is misconfigured.

The bug is that lan9645x_port_init now happens after npi_setup. This
will overwrite the REW_TAG_CFG settings on the npi port causing problems
with the standalone vlan tag leaking.

Fixes: bb7c9e32773b ("dsa: lan9645x: add frame preemption")