lan966x: Zero out short frames for APPL
There are cases when APPL sends a frame and the size of the frame is
only 42 bytes but the total size of the frame (IFH + frame) is bigger
than the minumum size of the frame(60 or 64 depending of the vlan,
without the FCS). In that case the frame was not padded with 0, so it
could be that garbage data will be send out. But this garbage data could
be some passwords. So make sure to zero out the frames correctly.

DBB-1003