lan969x: Fix calendar configuration
When configuring on EV23X71A to use only 1 SFP port then the traffic was
not working anymore on that port. If only port 24 or port 25 was enabled
in the device tree, then the traffic on that port was not passing
anymore. The same issue was seen on ports 26 and port port 27. The issue
seems to be with the calendar configuration as it was not set correct
because we need to set each index in the calendar to allocate a
bandwidth.

DBB-1013