arm: microchip: lan9645x: Add latencies for the Cu ports
Add latencies for the lan9645x evb board for the Cu ports. Without these
values if a 1m cable is connected back to back and run ptp4l it would
give us delay of ~600ns. With these values we get a value of 0 for 0.5m
cable.
The values were provided by the SQA.