net: hsr: populate skb_redundancy_info for HSR PTP frames
Populate skb_redundancy_info for HSR-tagged IEEE 1588  PTP frames in
the HSR forwarding path.

When creating an HSR-tagged frame in software, extract
redundancy-related metadata from the HSR header and store it in
skb_redundancy_info, including:
    - io_port: transmit direction and HSR leg
    - pathid: originating HSR leg
    - seqnr: HSR sequence number
    - ethertype and LSDU size

This metadata is only populated for HSR L2 PTP frames, and if tag
insertion is not offloaded.

For non-PTP frames, normal HSR forwarding behaviour is unchanged.

Signed-off-by: Daniel Machon <daniel.machon@microchip.com>